feat(rivm): Add rivm django app to populate db from file and serve graphql
This is Django project with pre build User app and default Django config.
To implement the rivm csv database population, the app inside this project
namingly rivm/rivm will be serving the GraphQL resources population during
application build.

This project comes with a build system and once that is run, and database is
populated it will not populate the database again given that the data source
is already read.